// libs/services/src/lib/models/LayoutActions.ts
export enum MenuState { 
    Open = 20,
    Closed = 21    
}
export enum MenuAction {  

    State_Open = 20,
    State_Close = 21,
    State_Toggle = 22,  
  
    Resize = 40,
    Resize_Small = 41,
    Resize_Medium = 42,
    Resize_Large  = 43,
  
    SwitchMode = 50,
    SwitchMode_Over = 51,
    SwitchMode_Push = 52,
    SwitchMode_Side = 53
  }
  